% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amstelgold}
\alias{amstelgold}
\title{Ride data for the Amstel Gold Race 2019}
\format{
A data frame with 69867 rows and 9 variables:
\describe{
\item{racetime}{time since the start of the race in seconds}
\item{athlete}{athlete identifier}
\item{altitude}{altitude measured by GPS in m}
\item{lat}{latitude measured by GPS in degree}
\item{lon}{longitude measured by GPS in degree}
\item{cadence}{pedalling cadence in 1/min}
\item{heartrate}{measured heart rate in bpm}
\item{temperature}{temperature in degree Celsius}
\item{power}{measured power in watts}
}
}
\source{
\url{https://strava.com/}
}
\usage{
amstelgold
}
\description{
A dataset containing GPS, power and heart rate data from three athletes
finishing the Amstel Gold Race 2019 under the top ten riders.
}
\keyword{datasets}
