% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gymnastics}
\alias{gymnastics}
\title{All gymnasts competing in the Rio 2016 Olympic Games.}
\format{
A data frame with 196 rows and 6 variables:
\describe{
\item{name}{name of the athlete}
\item{sex}{sex of the athlete}
\item{age}{age of the athlete in years}
\item{height}{body length in cm}
\item{weight}{bodyweight in kg}
\item{nation}{nationality of the athlete}
}
}
\usage{
gymnastics
}
\description{
A dataset containing information on athletes competing in the gymnastics
events at the Rio 2016 Olympic Games.
}
\keyword{datasets}
