% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympics}
\alias{olympics}
\title{All athletes competing in the Olympic Games}
\format{
A data frame with 271116 rows and 14 variables:
\describe{
\item{name}{name of the athlete}
\item{sex}{sex of the athlete}
\item{age}{age in years}
\item{height}{body height in cm}
\item{weight}{bodyweight in kg}
\item{nation}{nationality}
\item{noc}{National Olympic Commitee to compete for}
\item{games}{location and year of the Olympic Games}
\item{year}{year of the Olympic Games}
\item{season}{either \code{Summer} or \code{Winter}}
\item{city}{host city}
\item{sport}{type of sport}
\item{event}{event to participate}
\item{medal}{which medal was won}
}
}
\source{
\url{https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results}
}
\usage{
olympics
}
\description{
A dataset containing all athletes competing in any event of the modern
Olympic Games.
}
\keyword{datasets}
