% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rioathletics}
\alias{rioathletics}
\title{Some medal winners in Rio 2016 athletic competitions}
\format{
A data frame with 18 rows and 8 variables:
\describe{
\item{name}{name of the athlete}
\item{sex}{sex of the athlete}
\item{age}{age of the athlete in years}
\item{height}{body length in cm}
\item{weight}{bodyweight in kg}
\item{nation}{nationality of the athlete}
\item{event}{event in which the medal was won}
\item{medal}{type of medal won}
}
}
\usage{
rioathletics
}
\description{
A small example dataset containing information for medal winners in shot put,
high jump and 5000m running in the Rio 2016 Olympic Games.
}
\keyword{datasets}
