% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swimstars}
\alias{swimstars}
\title{Some of the best swimmers in history}
\format{
A data frame with 5 rows and 5 variables:
\describe{
\item{name}{name of the swimmer}
\item{sex}{sex of the swimmer}
\item{nation}{nationality of the swimmer}
\item{olympics}{number of Olympic Games participated}
\item{medals}{number of Olympic medals won}
}
}
\usage{
swimstars
}
\description{
A small example dataset containing information for some of the most famous
and successful swimmers in the world.
}
\keyword{datasets}
