% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rioathletics}
\alias{rioathletics}
\title{Contestants in some Rio 2016 Olympic athletic competitions}
\format{
A data frame with 233 rows and 8 variables:
\describe{
\item{name}{name of the athlete}
\item{sex}{sex of the athlete}
\item{age}{age of the athlete in years}
\item{height}{body length in cm}
\item{weight}{bodyweight in kg}
\item{nation}{nationality of the athlete}
\item{event}{event in which the medal was won}
\item{medal}{type of medal won}
}
}
\usage{
rioathletics
}
\description{
A dataset containing information for athletes competing in the shot put,
high jump and 5000m running events at the Rio 2016 Olympic Games.
}
\keyword{datasets}
